In support of #5806, changes to allow x64 Windows builds in the batlab. x64 Windows builds currently disabled, they can be enabled by editing remote_pre in the nmi build glue.  x64 Windows procd works, but daemon_core currently does not (socket passing is broken, not sure what else...)

+++ src/classad/classad/classad.h
@@ -541,7 +541,7 @@ class ClassAd : public ExprTree
 
         /** Return the number of attributes at the root level of this ClassAd.
          */
-        int size(void) const { return attrList.size(); }
+        int size(void) const { return (int)attrList.size(); }
 		//@}
 
 		void rehash(size_t s) { attrList.rehash(s);}
